refactor: move source directory '/src'
