+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.5",
   "description": "JavaScript library for npm package",
   "license": "MIT",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
     "test": "jest"
   },
